+++ packages/abc/tsconfig.lib.json
@@ -7,11 +7,5 @@
       "@delon/abc/*": ["../../dist/@delon/abc/*"],
       "@delon/*": ["../../dist/@delon/*"]
     }
-  },
-  "files": ["public_api.ts", "../typings.d.ts"],
-  "angularCompilerOptions": {
-    "annotateForClosureCompiler": true,
-    "strictMetadataEmit": false,
-    "skipTemplateCodegen": true
   }
 }
