+++ packages/auth/index.ts
@@ -1,1 +1,18 @@
-export * from './public_api';
+export * from './src/social/social.service';
+export * from './src/store/interface';
+export * from './src/store/local-storage.service';
+export * from './src/store/memory.service';
+export * from './src/store/session-storage.service';
+export * from './src/store/cookie-storage.service';
+export * from './src/token/base.interceptor';
+export * from './src/token/interface';
+export * from './src/token/token.service';
+export * from './src/token/jwt/jwt.helper';
+export * from './src/token/jwt/jwt.model';
+export * from './src/token/jwt/jwt.interceptor';
+export * from './src/token/jwt/jwt.guard';
+export * from './src/token/simple/simple.model';
+export * from './src/token/simple/simple.interceptor';
+export * from './src/token/simple/simple.guard';
+export * from './src/auth.config';
+export * from './src/auth.module';
