+++ packages/auth/ng-package.json
@@ -4,7 +4,7 @@
   "deleteDestPath": true,
   "lib": {
     "flatModuleFile": "auth",
-    "entryFile": "public_api.ts",
+    "entryFile": "index.ts",
     "umdModuleIds": {
       "@delon/util": "delon.util"
     }
