+++ packages/chart/tsconfig.lib.json
@@ -8,10 +8,5 @@
       "@delon/*": ["../../dist/@delon/*"]
     }
   },
-  "files": ["public_api.ts", "../typings.d.ts"],
-  "angularCompilerOptions": {
-    "annotateForClosureCompiler": true,
-    "strictMetadataEmit": false,
-    "skipTemplateCodegen": true
-  }
+  "files": ["public_api.ts", "../typings.d.ts"]
 }
