--- packages/util/decorator/factory.ts
@@ -1,35 +1,0 @@
-import { warn } from '@delon/util/other';
-import { NzSafeAny } from 'ng-zorro-antd/core/types';
-
-export function propDecoratorFactory<T, D>(
-  name: string,
-  fallback: (v: T, defaultValue: D) => D,
-  defaultValue: NzSafeAny,
-): (target: NzSafeAny, propName: string) => void {
-  function propDecorator(target: NzSafeAny, propName: string, originalDescriptor?: TypedPropertyDescriptor<NzSafeAny>): NzSafeAny {
-    const privatePropName = `$$__${propName}`;
-
-    if (Object.prototype.hasOwnProperty.call(target, privatePropName)) {
-      warn(`The prop "${privatePropName}" is already exist, it will be overrided by ${name} decorator.`);
-    }
-
-    Object.defineProperty(target, privatePropName, {
-      configurable: true,
-      writable: true,
-    });
-
-    return {
-      get(): string {
-        return originalDescriptor && originalDescriptor.get ? originalDescriptor.get.bind(this)() : this[privatePropName];
-      },
-      set(value: T): void {
-        if (originalDescriptor && originalDescriptor.set) {
-          originalDescriptor.set.bind(this)(fallback(value, defaultValue));
-        }
-        this[privatePropName] = fallback(value, defaultValue);
-      },
-    };
-  }
-
-  return propDecorator;
-}
