+++ packages/util/decorator/index.ts
@@ -1,3 +1,1 @@
-export * from './boolean-property';
-export * from './number-property';
-export * from './factory';
+export * from './convert';
