+++ packages/util/tsconfig.lib.json
@@ -7,6 +7,5 @@
       "@delon/util/*": ["../../dist/@delon/util/*"],
       "@delon/*": ["../../dist/@delon/*"]
     }
-  },
-  "files": ["index.ts", "../typings.d.ts"]
+  }
 }
